move gas_report_fuzz_invariant to integration
